+++ stroom-activity/stroom-activity-api/src/main/java/stroom/activity/api/ActivityService.java
@@ -19,6 +19,9 @@ package stroom.activity.api;
 import stroom.activity.shared.Activity;
 import stroom.activity.shared.ActivityValidationResult;
 import stroom.util.shared.ResultPage;
+import stroom.util.shared.filter.FilterFieldDefinition;
+
+import java.util.List;
 
 public interface ActivityService {
     Activity create();
@@ -29,7 +32,9 @@ public interface ActivityService {
 
     boolean delete(int id);
 
-    ResultPage<Activity> find(FindActivityCriteria criteria);
+    ResultPage<Activity> find(final String filter);
+
+    List<FilterFieldDefinition> listFieldDefinitions();
 
     ActivityValidationResult validate(Activity activity);
 }