+++ stroom-core-client/src/main/java/stroom/activity/client/ManageActivityViewImpl.java
@@ -16,7 +16,12 @@
 
 package stroom.activity.client;
 
+import stroom.activity.client.ManageActivityPresenter.ManageActivityView;
+import stroom.widget.dropdowntree.client.view.QuickFilter;
+import stroom.widget.layout.client.view.ResizeSimplePanel;
+
 import com.google.gwt.event.logical.shared.ValueChangeEvent;
+import com.google.gwt.safehtml.shared.SafeHtml;
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.uibinder.client.UiField;
 import com.google.gwt.uibinder.client.UiHandler;
@@ -24,9 +29,8 @@ import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 import com.gwtplatform.mvp.client.ViewWithUiHandlers;
-import stroom.activity.client.ManageActivityPresenter.ManageActivityView;
-import stroom.widget.dropdowntree.client.view.QuickFilter;
-import stroom.widget.layout.client.view.ResizeSimplePanel;
+
+import java.util.function.Supplier;
 
 
 public class ManageActivityViewImpl extends ViewWithUiHandlers<ManageActivityUiHandlers> implements ManageActivityView {
@@ -40,10 +44,15 @@ public class ManageActivityViewImpl extends ViewWithUiHandlers<ManageActivityUiH
     @UiField
     ResizeSimplePanel dataGrid;
 
+    private Supplier<SafeHtml> tooltipContentSupplier;
+
     @Inject
     public ManageActivityViewImpl(final EventBus eventBus, final Binder binder) {
         widget = binder.createAndBindUi(this);
+
+        nameFilter.registerPopupTextProvider(this::getTooltipContent);
     }
+
 
     @Override
     public void setInSlot(final Object slot, final Widget content) {
@@ -55,6 +64,15 @@ public class ManageActivityViewImpl extends ViewWithUiHandlers<ManageActivityUiH
     @UiHandler("nameFilter")
     void onFilterChange(final ValueChangeEvent<String> event) {
         getUiHandlers().changeNameFilter(nameFilter.getText());
+    }
+
+    @Override
+    public void setTooltipContentSupplier(final Supplier<SafeHtml> tooltipContentSupplier) {
+        this.tooltipContentSupplier = tooltipContentSupplier;
+    }
+
+    private SafeHtml getTooltipContent() {
+        return tooltipContentSupplier.get();
     }
 
     @Override