+++ stroom-importexport/stroom-importexport-impl/src/main/java/stroom/importexport/impl/DependencyService.java
@@ -5,5 +5,5 @@ import stroom.importexport.shared.DependencyCriteria;
 import stroom.util.shared.ResultPage;
 
 public interface DependencyService {
-    ResultPage<Dependency> getDependencies(DependencyCriteria criteria);
+    ResultPage<Dependency> getDependencies(final DependencyCriteria criteria);
 }