+++ stroom-importexport/stroom-importexport-impl/src/main/java/stroom/importexport/impl/DependencyServiceImpl.java
@@ -3,8 +3,15 @@ package stroom.importexport.impl;
 import stroom.docref.DocRef;
 import stroom.importexport.shared.Dependency;
 import stroom.importexport.shared.DependencyCriteria;
+import stroom.query.api.v2.ExpressionOperator;
+import stroom.query.api.v2.ExpressionOperator.Op;
 import stroom.task.api.TaskContext;
 import stroom.task.api.TaskContextFactory;
+import stroom.util.filter.FilterFieldMapper;
+import stroom.util.filter.FilterFieldMappers;
+import stroom.util.filter.QuickFilterPredicateFactory;
+import stroom.util.shared.CompareUtil;
+import stroom.util.shared.PageRequest;
 import stroom.util.shared.ResultPage;
 import stroom.util.shared.Sort;
 import stroom.util.shared.Sort.Direction;
@@ -13,12 +20,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
-import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
+import java.util.function.Function;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -27,6 +36,46 @@ public class DependencyServiceImpl implements DependencyService {
 
     private final ImportExportActionHandlers importExportActionHandlers;
     private final TaskContextFactory taskContextFactory;
+
+    private static final Comparator<Dependency> FROM_TYPE_COMPARATOR =
+            CompareUtil.getNullSafeCaseInsensitiveComparator(Dependency::getFrom, DocRef::getType);
+    private static final Comparator<Dependency> FROM_NAME_COMPARATOR =
+            CompareUtil.getNullSafeCaseInsensitiveComparator(Dependency::getFrom, DocRef::getName);
+    private static final Comparator<Dependency> FROM_UUID_COMPARATOR =
+            CompareUtil.getNullSafeCaseInsensitiveComparator(Dependency::getFrom, DocRef::getUuid);
+    private static final Comparator<Dependency> TO_TYPE_COMPARATOR =
+            CompareUtil.getNullSafeCaseInsensitiveComparator(Dependency::getTo, DocRef::getType);
+    private static final Comparator<Dependency> TO_NAME_COMPARATOR =
+            CompareUtil.getNullSafeCaseInsensitiveComparator(Dependency::getTo, DocRef::getName);
+    private static final Comparator<Dependency> TO_UUID_COMPARATOR =
+            CompareUtil.getNullSafeCaseInsensitiveComparator(Dependency::getTo, DocRef::getUuid);
+
+    private static final Map<String, Comparator<Dependency>> COMPARATOR_MAP = Map.of(
+            DependencyCriteria.FIELD_FROM_TYPE, FROM_TYPE_COMPARATOR,
+            DependencyCriteria.FIELD_FROM_NAME, FROM_NAME_COMPARATOR,
+            DependencyCriteria.FIELD_FROM_UUID, FROM_UUID_COMPARATOR,
+            DependencyCriteria.FIELD_TO_TYPE, TO_TYPE_COMPARATOR,
+            DependencyCriteria.FIELD_TO_NAME, TO_NAME_COMPARATOR,
+            DependencyCriteria.FIELD_TO_UUID, TO_UUID_COMPARATOR,
+            DependencyCriteria.FIELD_STATUS, Comparator.comparing(Dependency::isOk)
+    );
+
+    private static final Comparator<Dependency> DEFAULT_COMPARATOR = FROM_TYPE_COMPARATOR
+            .thenComparing(FROM_NAME_COMPARATOR)
+            .thenComparing(TO_TYPE_COMPARATOR)
+            .thenComparing(TO_NAME_COMPARATOR);
+
+    private static final ExpressionOperator SELECT_ALL_EXPRESSION_OP = new ExpressionOperator.Builder(Op.AND).build();
+
+    private static final FilterFieldMappers<Dependency> FIELD_MAPPERS = FilterFieldMappers.of(
+            FilterFieldMapper.of(DependencyCriteria.FIELD_DEF_FROM_TYPE, Dependency::getFrom, DocRef::getType),
+            FilterFieldMapper.of(DependencyCriteria.FIELD_DEF_FROM_NAME, Dependency::getFrom, DocRef::getName),
+            FilterFieldMapper.of(DependencyCriteria.FIELD_DEF_FROM_UUID, Dependency::getFrom, DocRef::getUuid),
+            FilterFieldMapper.of(DependencyCriteria.FIELD_DEF_TO_TYPE, Dependency::getTo, DocRef::getType),
+            FilterFieldMapper.of(DependencyCriteria.FIELD_DEF_TO_NAME, Dependency::getTo, DocRef::getName),
+            FilterFieldMapper.of(DependencyCriteria.FIELD_DEF_TO_UUID, Dependency::getTo, DocRef::getUuid),
+            FilterFieldMapper.of(DependencyCriteria.FIELD_DEF_STATUS, Dependency::isOk, bool -> bool ? "OK" : "Missing")
+    );
 
     @Inject
     public DependencyServiceImpl(final ImportExportActionHandlers importExportActionHandlers,
@@ -37,77 +86,189 @@ public class DependencyServiceImpl implements DependencyService {
 
     @Override
     public ResultPage<Dependency> getDependencies(final DependencyCriteria criteria) {
-        return taskContextFactory.contextResult("Get Dependencies", taskContext ->
-                getDependencies(criteria, taskContext)).get();
+        return taskContextFactory.contextResult(
+                "Get Dependencies",
+                taskContext -> {
+                    try {
+                        return getDependencies(criteria, taskContext);
+                    } catch (Exception e) {
+                        LOGGER.error("Error getting dependencies for criteria " + criteria, e);
+                        throw e;
+                    }
+                })
+                .get();
     }
 
-    private ResultPage<Dependency> getDependencies(final DependencyCriteria criteria, final TaskContext parentTaskContext) {
-        final Map<DocRef, Set<DocRef>> allDependencies = importExportActionHandlers
+    private ResultPage<Dependency> getDependencies(final DependencyCriteria criteria,
+                                                   final TaskContext parentTaskContext) {
+        // Build a map of deps (parent to children)
+        final Map<DocRef, Set<DocRef>> allDependencies = buildDependencyMap(parentTaskContext);
+
+        final Comparator<Dependency> sortListComparator = getDependencyComparator(criteria);
+
+        final Predicate<Dependency> filterPredicate = buildFilterPredicate(criteria);
+
+        // Flatten the dependency map
+        final List<Dependency> flatDependencies = buildFlatDependencies(
+                allDependencies,
+                sortListComparator,
+                filterPredicate);
+
+        return ResultPage.createPageLimitedList(
+                flatDependencies,
+                Optional.ofNullable(criteria)
+                        .map(DependencyCriteria::getPageRequest)
+                        .orElse(new PageRequest()));
+    }
+
+    private List<Dependency> buildFlatDependencies(final Map<DocRef, Set<DocRef>> allDependencies,
+                                                   final Comparator<Dependency> sortListComparator,
+                                                   final Predicate<Dependency> filterPredicate) {
+        return allDependencies.entrySet().stream()
+                .flatMap(entry -> {
+                    final DocRef parentDocRef = entry.getKey();
+                    final Set<DocRef> childDocRefs = entry.getValue();
+                    return childDocRefs.stream()
+                            .map(childDocRef -> new Dependency(
+                                    parentDocRef,
+                                    childDocRef,
+                                    allDependencies.containsKey(childDocRef)));
+                })
+                .filter(filterPredicate)
+                .sorted(sortListComparator)
+                .collect(Collectors.toList());
+    }
+
+    private Map<DocRef, Set<DocRef>> buildDependencyMap(final TaskContext parentTaskContext) {
+        return importExportActionHandlers
                 .getHandlers()
                 .values()
                 .parallelStream()
                 .map(handler ->
-                        taskContextFactory.contextResult(parentTaskContext, "Get " + handler.getType() + " dependencies", taskContext -> {
-                            Map<DocRef, Set<DocRef>> deps = null;
-                            try {
-                                deps = handler.getDependencies();
-                            } catch (final RuntimeException e) {
-                                LOGGER.error(e.getMessage(), e);
-                            }
-                            return deps;
-                        }).get())
+                        taskContextFactory.contextResult(
+                                parentTaskContext,
+                                "Get " + handler.getType() + " dependencies",
+                                taskContext -> {
+                                    Map<DocRef, Set<DocRef>> deps = null;
+                                    try {
+                                        deps = handler.getDependencies();
+                                    } catch (final RuntimeException e) {
+                                        LOGGER.error(e.getMessage(), e);
+                                    }
+                                    return deps;
+                                }).get())
                 .filter(Objects::nonNull)
                 .map(Map::entrySet)
                 .flatMap(Set::stream)
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (e1, e2) ->
-                        Stream.concat(e1.stream(), e2.stream()).collect(Collectors.toSet())));
+                .collect(Collectors.toMap(
+                        Map.Entry::getKey,
+                        Map.Entry::getValue,
+                        (e1, e2) ->
+                                Stream.concat(e1.stream(), e2.stream())
+                                        .collect(Collectors.toSet())));
+    }
 
-        final List<Dependency> dependencies = new ArrayList<>();
-        allDependencies.forEach((key, value) ->
-                value.forEach(to ->
-                        dependencies.add(new Dependency(key, to, allDependencies.containsKey(to)))));
+    private Predicate<Dependency> buildFilterPredicate(final DependencyCriteria criteria) {
+        return QuickFilterPredicateFactory.createFuzzyMatchPredicate(criteria.getPartialName(), FIELD_MAPPERS);
 
-        if (criteria.getSortList() == null || criteria.getSortList().size() == 0) {
-            criteria.addSort(new Sort(DependencyCriteria.FIELD_FROM, Direction.ASCENDING, true));
-            criteria.addSort(new Sort(DependencyCriteria.FIELD_TO, Direction.ASCENDING, true));
-            criteria.addSort(new Sort(DependencyCriteria.FIELD_STATUS, Direction.ASCENDING, true));
+//        final Predicate<Dependency> filterPredicate;
+//        if (criteria != null && criteria.getPartialName() != null) {
+//            final Predicate<DocRef> docRefPredicate =
+//                    DocRefPredicateFactory.createFuzzyMatchPredicate(
+//                            criteria.getPartialName(),
+//                            MatchMode.NAME_OR_TYPE);
+//
+//            filterPredicate = dep -> {
+//                if (dep == null) {
+//                    return false;
+//                } else {
+//                    // Match on any of {from,to} {name,type}
+//                    return docRefPredicate.test(dep.getFrom())
+//                            || docRefPredicate.test(dep.getTo());
+//                }
+//            };
+//        } else {
+//            filterPredicate = dep -> true;
+//        }
+//        return filterPredicate;
+    }
+
+    private boolean applyPredicate(final Dependency dependency,
+                                   final Function<Dependency, DocRef> docRefExtractor,
+                                   final Function<DocRef, String> valueExtractor,
+                                   final Predicate<String> stringPredicate) {
+        final boolean result;
+        if (dependency != null) {
+            DocRef docRef = docRefExtractor.apply(dependency);
+            if (docRef != null) {
+                String val = valueExtractor.apply(docRef);
+                if (val != null) {
+                    result = stringPredicate.test(val);
+                } else {
+                    // Null val
+                    result = false;
+                }
+            } else {
+                // Null docref
+                result = false;
+            }
+        } else {
+            // Null dep
+            result = false;
         }
+        return result;
+    }
 
-        final Comparator<DocRef> docRefComparator = getDocRefComparator();
-        dependencies.sort((o1, o2) -> {
-            int diff = 0;
+    private Comparator<Dependency> getDependencyComparator(final DependencyCriteria criteria) {
+        // Make the sort comparator base on the criteria sort list
+        final Comparator<Dependency> sortListComparator;
+        if (criteria != null
+                && criteria.getSortList() != null
+                && !criteria.getSortList().isEmpty()) {
+            sortListComparator = buildComparatorFromSortList(criteria);
+        } else {
+            sortListComparator = DEFAULT_COMPARATOR;
+        }
+        return sortListComparator;
+    }
 
-            for (final Sort sort : criteria.getSortList()) {
-                switch (sort.getField()) {
-                    case DependencyCriteria.FIELD_FROM:
-                        diff = docRefComparator.compare(o1.getFrom(), o2.getFrom());
-                        break;
-                    case DependencyCriteria.FIELD_TO:
-                        diff = docRefComparator.compare(o1.getTo(), o2.getTo());
-                        break;
-                    case DependencyCriteria.FIELD_STATUS:
-                        diff = Boolean.compare(o1.isOk(), o2.isOk());
-                        break;
-                }
+    private Comparator<Dependency> buildComparatorFromSortList(
+            final DependencyCriteria dependencyCriteria) {
 
-                if (Direction.DESCENDING.equals(sort.getDirection())) {
-                    diff = diff * -1;
-                }
+        if (dependencyCriteria != null && !dependencyCriteria.getSortList().isEmpty()) {
+            Comparator<Dependency> compositeComparator = null;
 
-                if (diff != 0) {
-                    return diff;
+            for (final Sort sort : dependencyCriteria.getSortList()) {
+                Comparator<Dependency> comparator = COMPARATOR_MAP.get(sort.getField());
+                if (comparator != null) {
+                    if (Direction.DESCENDING.equals(sort.getDirection())) {
+                        comparator = comparator.reversed();
+                    }
+                    compositeComparator = compositeComparator != null
+                            ? compositeComparator.thenComparing(comparator)
+                            : comparator;
                 }
             }
-
-            return diff;
-        });
-
-        return ResultPage.createPageLimitedList(dependencies, criteria.getPageRequest());
+            return compositeComparator != null
+                    ? compositeComparator
+                    : Comparator.comparing(dep -> 0);
+        } else {
+            // Unsorted
+            return Comparator.comparing(dep -> 0);
+        }
     }
 
-    private Comparator<DocRef> getDocRefComparator() {
-        return Comparator.comparing(DocRef::getType, Comparator.nullsFirst(Comparator.naturalOrder()))
-                .thenComparing(DocRef::getName, Comparator.nullsFirst(Comparator.naturalOrder()))
-                .thenComparing(DocRef::getUuid, Comparator.nullsFirst(Comparator.naturalOrder()));
-    }
+//    private static  Comparator<Dependency> getComparator(
+//            final Function<Dependency, DocRef> docRefExtractor,
+//            final Function<DocRef, String> valueExtractor) {
+//
+//        // Sort with nulls first but also handle deps with null docref
+//        return Comparator.comparing(
+//                docRefExtractor,
+//                Comparator.nullsFirst(
+//                        Comparator.comparing(
+//                                valueExtractor,
+//                                Comparator.nullsFirst(String.CASE_INSENSITIVE_ORDER))));
+//    }
+
 }